+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] 2023-11-03
+### Added
+- Refactor of how the stubs are processed
+- Addition of recursive property generation
+- Added a few flags to the `generate` command
+
+## [0.X.0] 2023-10-XX
+### Changed
+- Lots of small fixes
+
 ## [0.1.0] 2023-10-27
 ### Created
 - Initial setup of this library